requirements.txt
# Core
requests>=2.31.0
pillow>=10.0.0
pdf2image>=1.16.0
pymupdf>=1.23.0
markdown>=3.5.0
ollama>=0.1.0
python-dotenv>=1.0.0

# LLM Providers
anthropic>=0.25.0
openai>=1.0.0
google-generativeai>=0.3.0

# Configuration
pyyaml>=6.0.0

# Token counting
tiktoken>=0.6.0

# API (FastAPI)
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6

# Report generation
reportlab>=4.0.0

# Code Quality
black>=24.0.0
ruff>=0.4.0

# Testing
pytest>=8.0.0
pytest-cov>=4.0.0
